requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
openai==1.3.0
tinydb==4.8.0
pandas==2.1.0
plotly==5.17.0
passlib==1.7.4
python-dateutil==2.8.2
